Disabling optimization allows compilation to take only a few minutes. Can now save a basic, non-rotated image to a BMP file on disc. Bump to Kotlin 1.4.20
